Some lint fixes
