+++ config_test.go
@@ -1,10 +1,11 @@
-package hal
+package hal_test
 
 import (
 	"os"
 	"path/filepath"
 	"testing"
 
+	"github.com/dansimau/hal"
 	"gotest.tools/v3/assert"
 )
 
@@ -24,14 +25,14 @@ location:
   lng: -122.4194`
 
 		configPath := filepath.Join(tmpDir, "hal.yaml")
-		err := os.WriteFile(configPath, []byte(configContent), 0644)
+		err := os.WriteFile(configPath, []byte(configContent), 0o644)
 		assert.NilError(t, err)
 
 		// Change to temp directory
 		err = os.Chdir(tmpDir)
 		assert.NilError(t, err)
 
-		config, err := LoadConfig()
+		config, err := hal.LoadConfig()
 		assert.NilError(t, err)
 		assert.Equal(t, config.HomeAssistant.Host, "localhost:8123")
 		assert.Equal(t, config.HomeAssistant.Token, "test-token")
@@ -48,7 +49,7 @@ location:
 		err := os.Chdir(tmpDir)
 		assert.NilError(t, err)
 
-		_, err = LoadConfig()
+		_, err = hal.LoadConfig()
 		assert.ErrorContains(t, err, "no such file or directory")
 	})
 
@@ -59,137 +60,13 @@ location:
 
 		invalidContent := `invalid: yaml: content:`
 		configPath := filepath.Join(tmpDir, "hal.yaml")
-		err := os.WriteFile(configPath, []byte(invalidContent), 0644)
+		err := os.WriteFile(configPath, []byte(invalidContent), 0o644)
 		assert.NilError(t, err)
 
 		err = os.Chdir(tmpDir)
 		assert.NilError(t, err)
 
-		_, err = LoadConfig()
+		_, err = hal.LoadConfig()
 		assert.ErrorContains(t, err, "yaml")
-	})
-}
-
-func TestSearchParentsForFile(t *testing.T) {
-	t.Run("finds file in current directory", func(t *testing.T) {
-		tmpDir := t.TempDir()
-		testFile := "test.txt"
-		testPath := filepath.Join(tmpDir, testFile)
-
-		err := os.WriteFile(testPath, []byte("test"), 0644)
-		assert.NilError(t, err)
-
-		foundPath, err := searchParentsForFile(testFile, tmpDir)
-		assert.NilError(t, err)
-		assert.Equal(t, foundPath, testPath)
-	})
-
-	t.Run("finds file in parent directory", func(t *testing.T) {
-		tmpDir := t.TempDir()
-		subDir := filepath.Join(tmpDir, "subdir")
-		err := os.Mkdir(subDir, 0755)
-		assert.NilError(t, err)
-
-		testFile := "test.txt"
-		testPath := filepath.Join(tmpDir, testFile)
-
-		err = os.WriteFile(testPath, []byte("test"), 0644)
-		assert.NilError(t, err)
-
-		foundPath, err := searchParentsForFile(testFile, subDir)
-		assert.NilError(t, err)
-		assert.Equal(t, foundPath, testPath)
-	})
-
-	t.Run("returns empty string when file not found", func(t *testing.T) {
-		tmpDir := t.TempDir()
-		foundPath, err := searchParentsForFile("nonexistent.txt", tmpDir)
-		assert.NilError(t, err)
-		assert.Equal(t, foundPath, "")
-	})
-}
-
-func TestSearchParentsForFileFromCwd(t *testing.T) {
-	t.Run("searches from current working directory", func(t *testing.T) {
-		tmpDir := t.TempDir()
-		originalDir, _ := os.Getwd()
-		defer os.Chdir(originalDir)
-
-		testFile := "test.txt"
-		testPath := filepath.Join(tmpDir, testFile)
-
-		err := os.WriteFile(testPath, []byte("test"), 0644)
-		assert.NilError(t, err)
-
-		err = os.Chdir(tmpDir)
-		assert.NilError(t, err)
-
-		foundPath, err := searchParentsForFileFromCwd(testFile)
-		assert.NilError(t, err)
-		
-		// Resolve symlinks to handle macOS /private/var vs /var differences
-		resolvedFound, err := filepath.EvalSymlinks(foundPath)
-		assert.NilError(t, err)
-		resolvedExpected, err := filepath.EvalSymlinks(testPath)
-		assert.NilError(t, err)
-		
-		assert.Equal(t, resolvedFound, resolvedExpected)
-	})
-}
-
-func TestGetParents(t *testing.T) {
-	t.Run("returns correct parent paths", func(t *testing.T) {
-		if filepath.Separator == '\\' {
-			// Windows paths
-			paths := getParents("C:\\Users\\test\\project")
-			expected := []string{
-				"C:\\Users\\test\\project",
-				"C:\\Users\\test",
-				"C:\\Users",
-				"C:\\",
-				"/",
-			}
-			assert.DeepEqual(t, paths, expected)
-		} else {
-			// Unix paths
-			paths := getParents("/home/user/project")
-			expected := []string{
-				"/home/user/project",
-				"/home/user",
-				"/home",
-				"/",
-			}
-			assert.DeepEqual(t, paths, expected)
-		}
-	})
-
-	t.Run("handles root directory", func(t *testing.T) {
-		paths := getParents("/")
-		expected := []string{"/"}
-		assert.DeepEqual(t, paths, expected)
-	})
-}
-
-func TestFileExists(t *testing.T) {
-	t.Run("returns true for existing file", func(t *testing.T) {
-		tmpDir := t.TempDir()
-		testPath := filepath.Join(tmpDir, "test.txt")
-
-		err := os.WriteFile(testPath, []byte("test"), 0644)
-		assert.NilError(t, err)
-
-		exists := fileExists(testPath)
-		assert.Equal(t, exists, true)
-	})
-
-	t.Run("returns false for non-existent file", func(t *testing.T) {
-		exists := fileExists("/nonexistent/path/file.txt")
-		assert.Equal(t, exists, false)
-	})
-
-	t.Run("returns true for existing directory", func(t *testing.T) {
-		tmpDir := t.TempDir()
-		exists := fileExists(tmpDir)
-		assert.Equal(t, exists, true)
 	})
 }
