+++ entity_binary_sensor_test.go
