+++ entity_input_boolean_test.go
